Rename managed_user to managed_access_key
